refactor: split dev server editor UI modules
