+++ scripts/dev-server.ts
@@ -37,6 +37,7 @@ import {
   proseMirrorDocJsonToEditorCommands,
   textBodyToProseMirrorDocJson,
 } from "../packages/editor-core/src/index.js";
+import { generateDevEditorHtml } from "./dev-server-editor/template.js";
 import { unsafeScriptInputAssertion } from "./unsafe-type-assertion.js";
 
 const DEFAULT_PORT = 3000;
@@ -745,1660 +746,12 @@ function watchSourceFiles(onChange: () => void): void {
 // --- HTML template ---
 
 function generateHtml(slides: SlideSvg[], pptxName: string): string {
-  const thumbnailsHtml = generateThumbnailsHtml(slides);
-
-  const firstSvg = slides.length > 0 ? slides[0].svg : "<p>No slides</p>";
-
-  return `<!DOCTYPE html>
-<html lang="ja">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>pptx-glimpse dev - ${escapeHtml(pptxName)}</title>
-  <style>
-    * { margin: 0; padding: 0; box-sizing: border-box; }
-    body {
-      font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, sans-serif;
-      background: #1a1a2e;
-      color: #e0e0e0;
-    }
-    #header {
-      padding: 12px 20px;
-      background: #16213e;
-      display: flex;
-      align-items: center;
-      justify-content: space-between;
-    }
-    #header h1 { font-size: 14px; font-weight: 600; color: #a0a0c0; }
-    #status { font-size: 12px; color: #4caf50; }
-    #status.rendering { color: #ff9800; }
-    #status.error { color: #f44336; }
-    #main { display: flex; height: calc(100vh - 48px); }
-    #sidebar {
-      width: 180px;
-      overflow-y: auto;
-      background: #16213e;
-      padding: 8px;
-      border-right: 1px solid #2a2a4a;
-    }
-    #editor-panel {
-      width: 260px;
-      background: #111827;
-      border-left: 1px solid #2a2a4a;
-      padding: 12px;
-      display: flex;
-      flex-direction: column;
-      gap: 10px;
-    }
-    #editor-panel label {
-      display: flex;
-      flex-direction: column;
-      gap: 4px;
-      color: #a0a0c0;
-      font-size: 11px;
-      font-weight: 600;
-    }
-    #editor-panel select,
-    #editor-panel input {
-      width: 100%;
-      min-height: 32px;
-      border: 1px solid #334155;
-      border-radius: 4px;
-      background: #0f172a;
-      color: #e5e7eb;
-      padding: 6px 8px;
-      font: inherit;
-    }
-    #editor-panel button {
-      min-height: 32px;
-      border: 1px solid #475569;
-      border-radius: 4px;
-      background: #1f2937;
-      color: #f8fafc;
-      cursor: pointer;
-      font: inherit;
-      font-weight: 600;
-    }
-    #editor-panel button:hover:not(:disabled) { background: #334155; }
-    #editor-panel button:disabled {
-      opacity: 0.45;
-      cursor: not-allowed;
-    }
-    #editor-actions {
-      display: grid;
-      grid-template-columns: repeat(2, minmax(0, 1fr));
-      gap: 8px;
-    }
-    #save-button { grid-column: 1 / -1; }
-    #editor-message {
-      min-height: 18px;
-      color: #94a3b8;
-      font-size: 11px;
-      overflow-wrap: anywhere;
-    }
-    .thumbnail {
-      margin-bottom: 8px;
-      padding: 4px;
-      border: 2px solid transparent;
-      border-radius: 4px;
-      cursor: pointer;
-      background: #fff;
-    }
-    .thumbnail.active { border-color: #4472c4; }
-    .thumbnail:hover { border-color: #6090d0; }
-    .thumb-label {
-      display: flex;
-      align-items: center;
-      justify-content: space-between;
-      gap: 4px;
-      font-size: 10px;
-      color: #888;
-      padding: 2px 0;
-      background: #16213e;
-    }
-    .thumb-title { min-width: 0; overflow: hidden; text-overflow: ellipsis; white-space: nowrap; }
-    .thumb-actions { display: flex; gap: 3px; }
-    .thumb-action {
-      min-width: 22px;
-      min-height: 20px;
-      border: 1px solid #334155;
-      border-radius: 3px;
-      background: #0f172a;
-      color: #e5e7eb;
-      cursor: pointer;
-      font-size: 10px;
-      font-weight: 700;
-      line-height: 1;
-    }
-    .thumb-action:hover:not(:disabled) { background: #1d4ed8; }
-    .thumb-action:disabled { opacity: 0.4; cursor: not-allowed; }
-    .thumb-svg svg { width: 100%; height: auto; display: block; }
-    #viewer {
-      flex: 1;
-      display: flex;
-      align-items: center;
-      justify-content: center;
-      padding: 20px;
-      overflow: auto;
-    }
-    #slide-container {
-      background: #fff;
-      border-radius: 4px;
-      box-shadow: 0 4px 20px rgba(0, 0, 0, 0.3);
-      max-width: 100%;
-      max-height: 100%;
-      position: relative;
-    }
-    #slide-container svg { display: block; width: 100%; height: auto; }
-    #selection-overlay {
-      position: absolute;
-      inset: 0;
-      width: 100%;
-      height: 100%;
-      overflow: visible;
-      touch-action: none;
-      z-index: 2;
-    }
-    #text-editor-overlay {
-      position: absolute;
-      min-width: 40px;
-      min-height: 28px;
-      padding: 4px;
-      border: 1px solid #2563eb;
-      background: rgba(255, 255, 255, 0.96);
-      color: #111827;
-      z-index: 3;
-      overflow: hidden;
-      box-shadow: 0 0 0 1px rgba(37, 99, 235, 0.25);
-    }
-    .text-editor-paragraph {
-      min-height: 20px;
-      line-height: 1.25;
-      white-space: pre-wrap;
-      overflow-wrap: anywhere;
-    }
-    .text-editor-run {
-      outline: none;
-      white-space: pre-wrap;
-    }
-    .text-run-format-toolbar {
-      display: grid;
-      grid-template-columns: repeat(3, 24px) 44px 20px 28px 20px minmax(48px, 1fr) 20px;
-      gap: 3px;
-      align-items: center;
-      margin-bottom: 4px;
-    }
-    .text-run-format-toolbar button,
-    .text-run-format-toolbar input {
-      min-height: 22px;
-      border: 1px solid #94a3b8;
-      border-radius: 4px;
-      background: #f8fafc;
-      color: #0f172a;
-      font-size: 10px;
-      font-weight: 650;
-    }
-    .text-run-format-toolbar button {
-      padding: 0;
-      min-width: 0;
-    }
-    .text-run-format-toolbar button[aria-pressed="true"] {
-      background: #dbeafe;
-      border-color: #2563eb;
-    }
-    .text-run-format-toolbar input {
-      min-width: 0;
-      padding: 2px 4px;
-      font-weight: 500;
-    }
-    .text-run-format-toolbar input[type="color"] {
-      padding: 1px;
-    }
-    .text-run-format-toolbar :disabled {
-      opacity: 0.45;
-      cursor: not-allowed;
-    }
-    .text-editor-actions {
-      display: flex;
-      justify-content: flex-end;
-      gap: 4px;
-      margin-top: 4px;
-    }
-    .text-editor-actions button {
-      min-height: 24px;
-      padding: 2px 8px;
-      border: 1px solid #64748b;
-      border-radius: 4px;
-      background: #f8fafc;
-      color: #0f172a;
-      font-size: 11px;
-      font-weight: 600;
-    }
-    .shape-hit-area {
-      cursor: move;
-      fill: transparent;
-      pointer-events: all;
-    }
-    .selection-box {
-      fill: none;
-      stroke: #0f766e;
-      stroke-width: 1.5;
-      vector-effect: non-scaling-stroke;
-      pointer-events: none;
-    }
-    .selection-handle {
-      fill: #f8fafc;
-      stroke: #0f766e;
-      stroke-width: 1.5;
-      cursor: nwse-resize;
-      vector-effect: non-scaling-stroke;
-      pointer-events: all;
-    }
-    .selection-handle[data-handle="ne"],
-    .selection-handle[data-handle="sw"] {
-      cursor: nesw-resize;
-    }
-    #info {
-      padding: 4px 20px;
-      background: #16213e;
-      font-size: 11px;
-      color: #666;
-      text-align: center;
-    }
-  </style>
-</head>
-<body>
-  <div id="header">
-    <h1>pptx-glimpse dev &mdash; ${escapeHtml(pptxName)}</h1>
-    <span id="status">Connected</span>
-  </div>
-  <div id="main">
-    <div id="sidebar">${thumbnailsHtml}</div>
-    <div id="viewer">
-      <div id="slide-container">${firstSvg}</div>
-    </div>
-    <div id="editor-panel">
-      <label>Text run<select id="text-run-select"></select></label>
-      <label>Text<input id="text-run-input" type="text"></label>
-      <button id="apply-text-button" type="button">Apply</button>
-      <label>Image<input id="image-replacement-input" data-testid="image-replacement-input" type="file" disabled></label>
-      <div id="editor-actions">
-        <button id="add-text-box-button" type="button">Add text box</button>
-        <button id="delete-shape-button" type="button" disabled>Delete shape</button>
-        <button id="undo-button" type="button">Undo</button>
-        <button id="redo-button" type="button">Redo</button>
-        <button id="save-button" type="button">Save</button>
-      </div>
-      <div id="editor-message"></div>
-    </div>
-  </div>
-  <div id="info">Slide 1 / ${String(slides.length)}</div>
-  <script>
-    var currentIndex = 0;
-    var slideCount = ${String(slides.length)};
-    var slides = ${jsonForScript(slides)};
-    var editorHistory = { canUndo: false, canRedo: false, undoDepth: 0, redoDepth: 0 };
-    var textRunOptions = [];
-    var shapeOptions = [];
-    var selectedShapeKey = null;
-    var selectedShape = null;
-    var dragState = null;
-    var activeTextEditor = null;
-    var shapeRequestId = 0;
-    var EMU_PER_PIXEL = ${String(EMU_PER_INCH / DEFAULT_DPI)};
-
-    function selectSlide(index) {
-      if (activeTextEditor) {
-        commitTextEditor()
-          .then(function () {
-            selectSlide(index);
-          })
-          .catch(function () {
-            // Keep the editor open; commitTextEditor already reported the failure.
-          });
-        return;
-      }
-      var slideChanged = currentIndex !== index;
-      currentIndex = index;
-      shapeOptions = [];
-      textRunOptions = [];
-      selectedShape = null;
-      if (slideChanged) selectedShapeKey = null;
-      syncImageReplacementInput();
-      var thumbs = document.querySelectorAll(".thumbnail");
-      for (var i = 0; i < thumbs.length; i++) {
-        if (i === index) {
-          thumbs[i].classList.add("active");
-        } else {
-          thumbs[i].classList.remove("active");
-        }
-      }
-      document.getElementById("slide-container").innerHTML =
-        slides[index] ? slides[index].svg : "<p>No slides</p>";
-      var svg = document.querySelector("#slide-container svg");
-      if (svg) {
-        svg.removeAttribute("width");
-        svg.removeAttribute("height");
-        svg.style.width = "100%";
-        svg.style.height = "auto";
-      }
-      document.getElementById("info").textContent =
-        "Slide " + (index + 1) + " / " + slideCount;
-      renderSelectionOverlay();
-      loadShapeOptions(index + 1);
-    }
-
-    function renderThumbnails() {
-      var sidebar = document.getElementById("sidebar");
-      sidebar.innerHTML = slides
-        .map(function (s, i) {
-          return '<div class="thumbnail' + (i === currentIndex ? ' active' : '') +
-            '" data-index="' + i + '">' +
-            '<div class="thumb-label">' +
-            '<span class="thumb-title">Slide ' + s.slideNumber + '</span>' +
-            '<span class="thumb-actions">' +
-            '<button class="thumb-action" data-testid="duplicate-slide-' + i + '" data-action="duplicate" data-index="' + i + '" type="button" title="Duplicate slide">D</button>' +
-            '<button class="thumb-action" data-testid="delete-slide-' + i + '" data-action="delete" data-index="' + i + '" type="button" title="Delete slide"' +
-            (slideCount <= 1 ? ' disabled' : '') + '>X</button>' +
-            '</span>' +
-            '</div>' +
-            '<div class="thumb-svg">' + s.svg + '</div>' +
-            '</div>';
-        })
-        .join("");
-
-      var thumbs = document.querySelectorAll(".thumbnail");
-      for (var i = 0; i < thumbs.length; i++) {
-        (function (idx) {
-          thumbs[idx].addEventListener("click", function () {
-            selectSlide(idx);
-          });
-        })(i);
-      }
-
-      Array.prototype.forEach.call(sidebar.querySelectorAll("[data-action]"), function (button) {
-        button.addEventListener("mousedown", function (event) {
-          event.preventDefault();
-        });
-        button.addEventListener("click", function (event) {
-          event.preventDefault();
-          event.stopPropagation();
-          var index = Number(button.getAttribute("data-index") || "-1");
-          var action = button.getAttribute("data-action");
-          if (action === "duplicate") duplicateSlide(index);
-          if (action === "delete") deleteSlide(index);
-        });
-      });
-    }
-
-    function updateHistory(history) {
-      editorHistory = history || editorHistory;
-      document.getElementById("undo-button").disabled = !editorHistory.canUndo;
-      document.getElementById("redo-button").disabled = !editorHistory.canRedo;
-      updateSelectedShapeActions();
-    }
-
-    function updateSelectedShapeActions() {
-      document.getElementById("delete-shape-button").disabled =
-        activeTextEditor || !selectedShape || !selectedShape.editableDelete;
-    }
-
-    function shapeKey(shape) {
-      return shape && shape.handle ? handleKey(shape.handle) : "";
-    }
-
-    function handleKey(handle) {
-      return [
-        handle.partPath || "",
-        handle.nodeId || "",
-        handle.relationshipId || "",
-        handle.orderingSlot == null ? "" : String(handle.orderingSlot)
-      ].join("\\u0000");
-    }
-
-    function syncSelection(selection) {
-      selectedShapeKey = selection && selection.shapeHandle ? handleKey(selection.shapeHandle) : selectedShapeKey;
-      selectedShape = null;
-      if (selectedShapeKey) {
-        for (var i = 0; i < shapeOptions.length; i++) {
-          if (shapeKey(shapeOptions[i]) === selectedShapeKey && shapeOptions[i].bounds) {
-            selectedShape = cloneShape(shapeOptions[i]);
-            break;
-          }
-        }
-      }
-      if (!selectedShape) selectedShapeKey = null;
-      syncImageReplacementInput();
-      renderSelectionOverlay();
-      updateSelectedShapeActions();
-    }
-
-    function cloneShape(shape) {
-      return {
-        id: shape.id,
-        kind: shape.kind,
-        name: shape.name,
-        handle: shape.handle,
-        editableDelete: shape.editableDelete === true,
-        editableTransform: shape.editableTransform,
-        editableTextBody: shape.editableTextBody,
-        editableImageReplacement: shape.editableImageReplacement,
-        bounds: {
-          x: shape.bounds.x,
-          y: shape.bounds.y,
-          width: shape.bounds.width,
-          height: shape.bounds.height
-        }
-      };
-    }
-
-    function setEditorMessage(message, isError) {
-      var element = document.getElementById("editor-message");
-      element.textContent = message;
-      element.style.color = isError ? "#fca5a5" : "#94a3b8";
-    }
-
-    function loadShapeOptions(slideNumber) {
-      var requestId = ++shapeRequestId;
-      fetch("/api/editor/shapes?slide=" + slideNumber)
-        .then(function (res) {
-          if (!res.ok) throw new Error("shape request failed");
-          return res.json();
-        })
-        .then(function (data) {
-          if (requestId !== shapeRequestId || slideNumber !== currentIndex + 1) return;
-          shapeOptions = (data.shapes || []).filter(function (shape) {
-            return shape.handle && shape.bounds && (shape.editableTransform || shape.editableImageReplacement);
-          });
-          textRunOptions = [];
-          data.shapes.forEach(function (shape) {
-            (shape.textRuns || []).forEach(function (run, index) {
-              textRunOptions.push({
-                label: (shape.name || shape.id) + " / " + (index + 1),
-                text: run.text,
-                handle: run.handle
-              });
-            });
-          });
-          var select = document.getElementById("text-run-select");
-          select.innerHTML = textRunOptions
-            .map(function (option, index) {
-              return '<option value="' + index + '">' + escapeHtmlClient(option.label) + '</option>';
-            })
-            .join("");
-          select.disabled = textRunOptions.length === 0;
-          document.getElementById("apply-text-button").disabled = textRunOptions.length === 0;
-          syncTextInput();
-          syncSelection();
-          syncImageReplacementInput();
-        })
-        .catch(function (err) {
-          setEditorMessage(err.message, true);
-        });
-    }
-
-    function syncTextInput() {
-      var select = document.getElementById("text-run-select");
-      var input = document.getElementById("text-run-input");
-      var option = textRunOptions[Number(select.value || "0")];
-      input.value = option ? option.text : "";
-      input.disabled = !option;
-    }
-
-    function applyEditorResponse(data, preferredIndex) {
-      slides = data.slides || slides;
-      slideCount = slides.length;
-      updateHistory(data.history);
-      if (data.selection && data.selection.shapeHandle) {
-        selectedShapeKey = handleKey(data.selection.shapeHandle);
-      }
-      renderThumbnails();
-      var nextIndex = preferredIndex == null ? currentIndex : preferredIndex;
-      selectSlide(Math.min(Math.max(nextIndex, 0), Math.max(slides.length - 1, 0)));
-    }
-
-    function renderSelectionOverlay() {
-      var container = document.getElementById("slide-container");
-      var renderedSvg = container.querySelector("svg:not(#selection-overlay)");
-      var existing = document.getElementById("selection-overlay");
-      if (existing) existing.remove();
-      if (!renderedSvg) return;
-
-      var viewBox = renderedSvg.getAttribute("viewBox") || "0 0 960 540";
-      var overlay = document.createElementNS("http://www.w3.org/2000/svg", "svg");
-      overlay.setAttribute("id", "selection-overlay");
-      overlay.setAttribute("viewBox", viewBox);
-      overlay.setAttribute("data-testid", "selection-overlay");
-
-      shapeOptions.forEach(function (shape, index) {
-        if (!shape.bounds) return;
-        var hit = document.createElementNS("http://www.w3.org/2000/svg", "rect");
-        hit.setAttribute("class", "shape-hit-area");
-        hit.setAttribute("data-testid", "shape-hit-area");
-        hit.setAttribute("data-shape-index", String(index));
-        setRectAttributes(hit, shape.bounds);
-        hit.addEventListener("pointerdown", function (event) {
-          selectShape(shape, event);
-        });
-        hit.addEventListener("mousedown", function (event) {
-          if (activeTextEditor) return;
-          if (event.detail >= 2 && shape.editableTextBody) {
-            event.preventDefault();
-            openTextEditor(shape);
-          }
-        });
-        hit.addEventListener("dblclick", function (event) {
-          event.preventDefault();
-          event.stopPropagation();
-          openTextEditor(shape);
-        });
-        overlay.appendChild(hit);
-      });
-
-      if (selectedShape && selectedShape.bounds) {
-        var box = document.createElementNS("http://www.w3.org/2000/svg", "rect");
-        box.setAttribute("class", "selection-box");
-        box.setAttribute("data-testid", "selection-box");
-        setRectAttributes(box, selectedShape.bounds);
-        overlay.appendChild(box);
-
-        if (selectedShape.editableTransform) {
-          ["nw", "ne", "sw", "se"].forEach(function (handle) {
-            var point = handlePoint(selectedShape.bounds, handle);
-            var rect = document.createElementNS("http://www.w3.org/2000/svg", "rect");
-            rect.setAttribute("class", "selection-handle");
-            rect.setAttribute("data-testid", "selection-handle-" + handle);
-            rect.setAttribute("data-handle", handle);
-            rect.setAttribute("x", String(point.x - 4));
-            rect.setAttribute("y", String(point.y - 4));
-            rect.setAttribute("width", "8");
-            rect.setAttribute("height", "8");
-            rect.addEventListener("pointerdown", function (event) {
-              event.preventDefault();
-              event.stopPropagation();
-              beginDrag("resize", handle, event);
-            });
-            overlay.appendChild(rect);
-          });
-        }
-      }
-
-      container.appendChild(overlay);
-      positionActiveTextEditor();
-      updateSelectedShapeActions();
-    }
-
-    function setRectAttributes(rect, bounds) {
-      rect.setAttribute("x", String(bounds.x));
-      rect.setAttribute("y", String(bounds.y));
-      rect.setAttribute("width", String(bounds.width));
-      rect.setAttribute("height", String(bounds.height));
-    }
-
-    function handlePoint(bounds, handle) {
-      var right = bounds.x + bounds.width;
-      var bottom = bounds.y + bounds.height;
-      if (handle === "nw") return { x: bounds.x, y: bounds.y };
-      if (handle === "ne") return { x: right, y: bounds.y };
-      if (handle === "sw") return { x: bounds.x, y: bottom };
-      return { x: right, y: bottom };
-    }
-
-    function selectShape(shape, event) {
-      if (activeTextEditor) {
-        commitTextEditor()
-          .then(function () {
-            selectShapeAfterCommit(shape);
-          })
-          .catch(function () {});
-        return;
-      }
-      selectedShapeKey = shapeKey(shape);
-      selectedShape = cloneShape(shape);
-      syncImageReplacementInput();
-      if (shape.editableTransform) beginDrag("move", null, event);
-      window.setTimeout(renderSelectionOverlay, 0);
-      postJson("/api/editor/select", { handle: shape.handle })
-        .then(function (data) {
-          updateHistory(data.history);
-          if (data.selection && data.selection.shapeHandle) {
-            selectedShapeKey = handleKey(data.selection.shapeHandle);
-          }
-          syncImageReplacementInput();
-          renderSelectionOverlay();
-        })
-        .catch(function (err) {
-          setEditorMessage(err.message, true);
-        });
-    }
-
-    function selectShapeAfterCommit(shape) {
-      selectedShapeKey = shapeKey(shape);
-      selectedShape = cloneShape(shape);
-      syncImageReplacementInput();
-      renderSelectionOverlay();
-      postJson("/api/editor/select", { handle: shape.handle })
-        .then(function (data) {
-          updateHistory(data.history);
-          if (data.selection && data.selection.shapeHandle) {
-            selectedShapeKey = handleKey(data.selection.shapeHandle);
-          }
-          syncImageReplacementInput();
-          renderSelectionOverlay();
-        })
-        .catch(function (err) {
-          setEditorMessage(err.message, true);
-        });
-    }
-
-    function openTextEditor(shape) {
-      if (!shape || !shape.handle || !shape.bounds || !shape.editableTextBody) return;
-      if (activeTextEditor) return;
-      if (dragState) {
-        dragState = null;
-        detachDragListeners();
-      }
-      closeTextEditor();
-      selectedShapeKey = shapeKey(shape);
-      selectedShape = cloneShape(shape);
-      renderSelectionOverlay();
-
-      var container = document.getElementById("slide-container");
-      var overlay = document.createElement("div");
-      overlay.id = "text-editor-overlay";
-      overlay.setAttribute("data-testid", "text-editor-overlay");
-      overlay.dataset.shapeKey = selectedShapeKey || "";
-      overlay.appendChild(createTextRunFormatToolbar());
-      overlay.appendChild(createTextEditorContent(shape.editableTextBody.docJson));
-
-      var actions = document.createElement("div");
-      actions.className = "text-editor-actions";
-      var done = document.createElement("button");
-      done.type = "button";
-      done.textContent = "Done";
-      done.setAttribute("data-testid", "text-editor-done");
-      done.addEventListener("click", function () {
-        commitTextEditor().catch(function () {});
-      });
-      actions.appendChild(done);
-      overlay.appendChild(actions);
-
-      overlay.addEventListener("keydown", function (event) {
-        if (event.isComposing || event.keyCode === 229) return;
-        if (event.key === "Enter" && !event.shiftKey) {
-          event.preventDefault();
-          commitTextEditor().catch(function () {});
-        }
-        if (event.key === "Escape") {
-          event.preventDefault();
-          closeTextEditor();
-        }
-      });
-      overlay.addEventListener("focusout", function (event) {
-        window.setTimeout(function () {
-          if (!activeTextEditor) return;
-          if (event.relatedTarget && activeTextEditor.element.contains(event.relatedTarget)) return;
-          commitTextEditor().catch(function () {});
-        }, 0);
-      });
-
-      container.appendChild(overlay);
-      activeTextEditor = {
-        element: overlay,
-        shape: cloneShape(shape),
-        originalDocJson: shape.editableTextBody.docJson,
-        selectedRunElement: null,
-        committing: false,
-        commitPromise: null
-      };
-      updateSelectedShapeActions();
-      positionActiveTextEditor();
-      var firstRun = overlay.querySelector(".text-editor-run");
-      if (firstRun) {
-        setActiveTextRunElement(firstRun);
-        firstRun.focus();
-        selectElementContents(firstRun);
-      }
-    }
-
-    function createTextRunFormatToolbar() {
-      var toolbar = document.createElement("div");
-      toolbar.className = "text-run-format-toolbar";
-      toolbar.setAttribute("data-testid", "text-run-format-toolbar");
-
-      [["bold", "B"], ["italic", "I"], ["underline", "U"]].forEach(function (item) {
-        var button = document.createElement("button");
-        button.type = "button";
-        button.textContent = item[1];
-        button.dataset.property = item[0];
-        button.setAttribute("data-testid", "text-run-format-" + item[0]);
-        button.setAttribute("aria-pressed", "false");
-        button.addEventListener("click", function () {
-          toggleActiveTextRunBooleanProperty(item[0]);
-        });
-        toolbar.appendChild(button);
-      });
-
-      var size = document.createElement("input");
-      size.type = "number";
-      size.min = "1";
-      size.step = "0.5";
-      size.placeholder = "pt";
-      size.setAttribute("data-testid", "text-run-format-font-size");
-      size.addEventListener("change", function () {
-        var value = Number(size.value);
-        if (size.value.trim() === "") {
-          applyActiveTextRunPropertyClear(["fontSize"]);
-        } else if (Number.isFinite(value) && value > 0) {
-          applyActiveTextRunPropertySet({ fontSize: value });
-        } else {
-          setEditorMessage("font size must be positive", true);
-        }
-      });
-      toolbar.appendChild(size);
-
-      var clearSize = clearButton("font-size", function () {
-        applyActiveTextRunPropertyClear(["fontSize"]);
-      });
-      toolbar.appendChild(clearSize);
-
-      var color = document.createElement("input");
-      color.type = "color";
-      color.setAttribute("data-testid", "text-run-format-color");
-      color.addEventListener("change", function () {
-        applyActiveTextRunPropertySet({
-          color: { kind: "srgb", hex: color.value.replace(/^#/, "").toUpperCase() }
-        });
-      });
-      toolbar.appendChild(color);
-
-      var clearColor = clearButton("color", function () {
-        applyActiveTextRunPropertyClear(["color"]);
-      });
-      toolbar.appendChild(clearColor);
-
-      var typeface = document.createElement("input");
-      typeface.type = "text";
-      typeface.placeholder = "Font";
-      typeface.setAttribute("data-testid", "text-run-format-typeface");
-      typeface.addEventListener("keydown", function (event) {
-        if (event.key === "Enter") {
-          event.preventDefault();
-          applyTypefaceInput(typeface);
-        }
-      });
-      typeface.addEventListener("change", function () {
-        applyTypefaceInput(typeface);
-      });
-      toolbar.appendChild(typeface);
-
-      var clearTypeface = clearButton("typeface", function () {
-        applyActiveTextRunPropertyClear(["typeface"]);
-      });
-      toolbar.appendChild(clearTypeface);
-
-      window.setTimeout(refreshTextRunFormatToolbar, 0);
-      return toolbar;
-    }
-
-    function clearButton(name, onClick) {
-      var button = document.createElement("button");
-      button.type = "button";
-      button.textContent = "x";
-      button.setAttribute("data-testid", "text-run-format-clear-" + name);
-      button.addEventListener("click", onClick);
-      return button;
-    }
-
-    function applyTypefaceInput(input) {
-      var value = input.value.trim();
-      if (value.length === 0) {
-        applyActiveTextRunPropertyClear(["typeface"]);
-        return;
-      }
-      applyActiveTextRunPropertySet({ typeface: value });
-    }
-
-    function setActiveTextRunElement(element) {
-      if (!activeTextEditor) return;
-      activeTextEditor.selectedRunElement = element;
-      refreshTextRunFormatToolbar();
-    }
-
-    function activeTextRunInfo() {
-      if (!activeTextEditor || !activeTextEditor.selectedRunElement) return null;
-      var element = activeTextEditor.selectedRunElement;
-      var paragraphIndex = Number(element.dataset.paragraphIndex || "-1");
-      var runIndex = Number(element.dataset.runIndex || "-1");
-      var paragraph = (activeTextEditor.originalDocJson.content || [])[paragraphIndex];
-      var textNode = paragraph && (paragraph.content || [])[runIndex];
-      var mark = textNode && (textNode.marks || []).find(function (candidate) {
-        return candidate.type === "pptxRun";
-      });
-      var attrs = mark && mark.attrs ? mark.attrs : {};
-      return {
-        handle: attrs.handle || null,
-        properties: attrs.properties || {}
-      };
-    }
-
-    function refreshTextRunFormatToolbar() {
-      if (!activeTextEditor) return;
-      var toolbar = activeTextEditor.element.querySelector('[data-testid="text-run-format-toolbar"]');
-      if (!toolbar) return;
-      var info = activeTextRunInfo();
-      var properties = info && info.properties ? info.properties : {};
-      var disabled = !info || !info.handle;
-      setToolbarControl(toolbar, "bold", disabled, Boolean(properties.bold));
-      setToolbarControl(toolbar, "italic", disabled, Boolean(properties.italic));
-      setToolbarControl(toolbar, "underline", disabled, Boolean(properties.underline));
-      var size = toolbar.querySelector('[data-testid="text-run-format-font-size"]');
-      if (size) {
-        size.disabled = disabled;
-        size.value = properties.fontSize == null ? "" : String(properties.fontSize);
-      }
-      var color = toolbar.querySelector('[data-testid="text-run-format-color"]');
-      if (color) {
-        color.disabled = disabled;
-        color.value =
-          properties.color && properties.color.kind === "srgb" && typeof properties.color.hex === "string"
-            ? "#" + properties.color.hex
-            : "#000000";
-      }
-      var typeface = toolbar.querySelector('[data-testid="text-run-format-typeface"]');
-      if (typeface) {
-        typeface.disabled = disabled;
-        typeface.value = typeof properties.typeface === "string" ? properties.typeface : "";
-      }
-      Array.prototype.forEach.call(toolbar.querySelectorAll('[data-testid^="text-run-format-clear-"]'), function (control) {
-        control.disabled = disabled;
-      });
-    }
-
-    function setToolbarControl(toolbar, property, disabled, pressed) {
-      var control = toolbar.querySelector('[data-testid="text-run-format-' + property + '"]');
-      if (!control) return;
-      control.disabled = disabled;
-      control.setAttribute("aria-pressed", pressed ? "true" : "false");
-    }
-
-    function toggleActiveTextRunBooleanProperty(property) {
-      var info = activeTextRunInfo();
-      if (!info || !info.handle) {
-        setEditorMessage("No editable text run selected", true);
-        return;
-      }
-      if ((info.properties || {})[property] === true) {
-        applyActiveTextRunPropertyClear([property]);
-        return;
-      }
-      var properties = {};
-      properties[property] = true;
-      applyActiveTextRunPropertySet(properties);
-    }
-
-    function applyActiveTextRunPropertySet(properties) {
-      var info = activeTextRunInfo();
-      if (!info || !info.handle) {
-        setEditorMessage("No editable text run selected", true);
-        return;
-      }
-      applyActiveTextRunPropertyCommand({
-        kind: "setTextRunProperties",
-        handle: info.handle,
-        properties: properties
-      });
-    }
-
-    function applyActiveTextRunPropertyClear(properties) {
-      var info = activeTextRunInfo();
-      if (!info || !info.handle) {
-        setEditorMessage("No editable text run selected", true);
-        return;
-      }
-      applyActiveTextRunPropertyCommand({
-        kind: "clearTextRunProperties",
-        handle: info.handle,
-        properties: properties
-      });
-    }
-
-    function applyActiveTextRunPropertyCommand(command) {
-      postJson("/api/editor/command", { command: command })
-        .then(function (data) {
-          applyEditorResponseBehindTextEditor(data);
-          patchActiveTextRunProperties(command);
-          setEditorMessage("Applied", false);
-        })
-        .catch(function (err) {
-          setEditorMessage(err.message, true);
-        });
-    }
-
-    function applyEditorResponseBehindTextEditor(data) {
-      slides = data.slides || slides;
-      slideCount = slides.length;
-      updateHistory(data.history);
-      if (data.selection && data.selection.shapeHandle) {
-        selectedShapeKey = handleKey(data.selection.shapeHandle);
-      }
-      renderThumbnails();
-      replaceCurrentRenderedSvg();
-      renderSelectionOverlay();
-      loadShapeOptions(currentIndex + 1);
-    }
-
-    function replaceCurrentRenderedSvg() {
-      var container = document.getElementById("slide-container");
-      var previous = container.querySelector("svg:not(#selection-overlay)");
-      var slide = slides[currentIndex];
-      if (!slide) return;
-      var wrapper = document.createElement("div");
-      wrapper.innerHTML = slide.svg;
-      var next = wrapper.querySelector("svg");
-      if (!next) return;
-      next.removeAttribute("width");
-      next.removeAttribute("height");
-      next.style.width = "100%";
-      next.style.height = "auto";
-      if (previous) {
-        previous.replaceWith(next);
-      } else {
-        container.insertBefore(next, container.firstChild);
-      }
-    }
-
-    function patchActiveTextRunProperties(command) {
-      var textNode = activeTextRunTextNode();
-      if (!textNode) return;
-      var mark = (textNode.marks || []).find(function (candidate) {
-        return candidate.type === "pptxRun";
-      });
-      if (!mark) return;
-      var attrs = mark.attrs || {};
-      var properties = { ...(attrs.properties || {}) };
-      if (command.kind === "setTextRunProperties") {
-        Object.keys(command.properties || {}).forEach(function (property) {
-          properties[property] = command.properties[property];
-        });
-      }
-      if (command.kind === "clearTextRunProperties") {
-        (command.properties || []).forEach(function (property) {
-          delete properties[property];
-        });
-      }
-      attrs.properties = Object.keys(properties).length > 0 ? properties : null;
-      mark.attrs = attrs;
-      refreshTextRunFormatToolbar();
-      syncActiveTextRunStyle(properties);
-    }
-
-    function activeTextRunTextNode() {
-      if (!activeTextEditor || !activeTextEditor.selectedRunElement) return null;
-      var element = activeTextEditor.selectedRunElement;
-      var paragraphIndex = Number(element.dataset.paragraphIndex || "-1");
-      var runIndex = Number(element.dataset.runIndex || "-1");
-      var paragraph = (activeTextEditor.originalDocJson.content || [])[paragraphIndex];
-      return paragraph && (paragraph.content || [])[runIndex] ? paragraph.content[runIndex] : null;
-    }
-
-    function syncActiveTextRunStyle(properties) {
-      if (!activeTextEditor || !activeTextEditor.selectedRunElement) return;
-      var run = activeTextEditor.selectedRunElement;
-      run.style.fontWeight = properties.bold === true ? "700" : "";
-      run.style.fontStyle = properties.italic === true ? "italic" : "";
-      run.style.textDecoration = properties.underline === true ? "underline" : "";
-      run.style.fontSize = properties.fontSize != null ? String(properties.fontSize) + "pt" : "";
-      run.style.fontFamily = properties.typeface
-        ? '"' + String(properties.typeface).replace(/"/g, "") + '"'
-        : "";
-      run.style.color =
-        properties.color && properties.color.kind === "srgb" && typeof properties.color.hex === "string"
-          ? "#" + properties.color.hex
-          : "";
-    }
-
-    function createTextEditorContent(docJson) {
-      var body = document.createElement("div");
-      body.setAttribute("data-testid", "text-editor-content");
-      (docJson.content || []).forEach(function (paragraph, paragraphIndex) {
-        var paragraphElement = document.createElement("div");
-        paragraphElement.className = "text-editor-paragraph";
-        paragraphElement.dataset.paragraphIndex = String(paragraphIndex);
-        (paragraph.content || []).forEach(function (textNode, runIndex) {
-          var run = document.createElement("span");
-          run.className = "text-editor-run";
-          run.setAttribute("data-testid", "text-editor-run");
-          run.contentEditable = "true";
-          run.dataset.paragraphIndex = String(paragraphIndex);
-          run.dataset.runIndex = String(runIndex);
-          run.textContent = textNode.text || "";
-          applyTextRunStyle(run, textNode);
-          run.addEventListener("focus", function () {
-            setActiveTextRunElement(run);
-          });
-          run.addEventListener("pointerdown", function () {
-            setActiveTextRunElement(run);
-          });
-          run.addEventListener("beforeinput", function (event) {
-            if (event.inputType === "insertParagraph" || event.inputType === "insertLineBreak") {
-              event.preventDefault();
-              commitTextEditor().catch(function () {});
-            }
-          });
-          run.addEventListener("paste", function (event) {
-            event.preventDefault();
-            var text = (event.clipboardData ? event.clipboardData.getData("text/plain") : "")
-              .replace(/\\r?\\n/g, " ");
-            document.execCommand("insertText", false, text);
-          });
-          paragraphElement.appendChild(run);
-        });
-        body.appendChild(paragraphElement);
-      });
-      return body;
-    }
-
-    function applyTextRunStyle(run, textNode) {
-      var mark = (textNode.marks || []).find(function (candidate) {
-        return candidate.type === "pptxRun";
-      });
-      var properties = mark && mark.attrs && mark.attrs.properties ? mark.attrs.properties : {};
-      if (properties.bold === true) run.style.fontWeight = "700";
-      if (properties.italic === true) run.style.fontStyle = "italic";
-      if (properties.underline === true) run.style.textDecoration = "underline";
-      if (properties.fontSize != null) run.style.fontSize = String(properties.fontSize) + "pt";
-      if (properties.typeface) run.style.fontFamily = '"' + String(properties.typeface).replace(/"/g, "") + '"';
-      if (properties.color && properties.color.kind === "srgb" && typeof properties.color.hex === "string") {
-        run.style.color = "#" + properties.color.hex;
-      }
-    }
-
-    function selectElementContents(element) {
-      var range = document.createRange();
-      range.selectNodeContents(element);
-      var selection = window.getSelection();
-      if (!selection) return;
-      selection.removeAllRanges();
-      selection.addRange(range);
-    }
-
-    function positionActiveTextEditor() {
-      if (!activeTextEditor || !activeTextEditor.shape || !activeTextEditor.shape.bounds) return;
-      var container = document.getElementById("slide-container");
-      var renderedSvg = container.querySelector("svg:not(#selection-overlay)");
-      if (!renderedSvg) return;
-      var svgRect = renderedSvg.getBoundingClientRect();
-      var containerRect = container.getBoundingClientRect();
-      var viewBox = parseViewBox(renderedSvg.getAttribute("viewBox") || "0 0 960 540");
-      var bounds = activeTextEditor.shape.bounds;
-      var left = svgRect.left - containerRect.left + ((bounds.x - viewBox.x) / viewBox.width) * svgRect.width;
-      var top = svgRect.top - containerRect.top + ((bounds.y - viewBox.y) / viewBox.height) * svgRect.height;
-      var width = (bounds.width / viewBox.width) * svgRect.width;
-      var height = (bounds.height / viewBox.height) * svgRect.height;
-      activeTextEditor.element.style.left = left + "px";
-      activeTextEditor.element.style.top = top + "px";
-      activeTextEditor.element.style.width = width + "px";
-      activeTextEditor.element.style.height = height + "px";
-    }
-
-    function parseViewBox(value) {
-      var parts = String(value).trim().split(/\\s+/).map(Number);
-      return {
-        x: Number.isFinite(parts[0]) ? parts[0] : 0,
-        y: Number.isFinite(parts[1]) ? parts[1] : 0,
-        width: Number.isFinite(parts[2]) && parts[2] > 0 ? parts[2] : 960,
-        height: Number.isFinite(parts[3]) && parts[3] > 0 ? parts[3] : 540
-      };
-    }
-
-    function textEditorDocJson() {
-      if (!activeTextEditor) return null;
-      var original = activeTextEditor.originalDocJson;
-      var paragraphs = [];
-      (original.content || []).forEach(function (paragraph, paragraphIndex) {
-        var paragraphElement = activeTextEditor.element.querySelector(
-          '.text-editor-paragraph[data-paragraph-index="' + paragraphIndex + '"]'
-        );
-        var content = [];
-        var emptyRunCount = 0;
-        (paragraph.content || []).forEach(function (textNode, runIndex) {
-          var runElement = paragraphElement
-            ? paragraphElement.querySelector('.text-editor-run[data-run-index="' + runIndex + '"]')
-            : null;
-          var text = runElement ? runElement.textContent || "" : textNode.text || "";
-          if (text.length === 0) {
-            emptyRunCount += 1;
-            return;
-          }
-          content.push({
-            type: "text",
-            text: text,
-            marks: textNode.marks || []
-          });
-        });
-        if (emptyRunCount > 0 && content.length > 0) {
-          throw new Error("Clearing individual runs in multi-run text is unsupported.");
-        }
-        paragraphs.push({
-          type: "paragraph",
-          attrs: paragraph.attrs || {},
-          content: content
-        });
-      });
-      return { type: "doc", content: paragraphs };
-    }
-
-    function commitTextEditor() {
-      if (!activeTextEditor) return Promise.resolve();
-      var editor = activeTextEditor;
-      if (editor.committing) return editor.commitPromise || Promise.resolve();
-      var docJson = textEditorDocJson();
-      if (!docJson) return Promise.resolve();
-      editor.committing = true;
-      editor.commitPromise = postJson("/api/editor/text-body", {
-        handle: editor.shape.handle,
-        docJson: docJson
-      })
-        .then(function (data) {
-          closeTextEditor(editor);
-          applyEditorResponse(data);
-          setEditorMessage("Applied", false);
-        })
-        .catch(function (err) {
-          editor.committing = false;
-          editor.commitPromise = null;
-          setEditorMessage(err.message, true);
-          throw err;
-        });
-      return editor.commitPromise;
-    }
-
-    function closeTextEditor(editor) {
-      var target = editor || activeTextEditor;
-      if (!target) return;
-      if (editor && activeTextEditor !== editor) return;
-      if (!editor && target.committing) return;
-      target.element.remove();
-      if (activeTextEditor === target) activeTextEditor = null;
-      updateSelectedShapeActions();
-    }
-
-    function beginDrag(kind, handle, event) {
-      if (!selectedShape || !selectedShape.bounds) return;
-      var overlay = document.getElementById("selection-overlay");
-      if (!overlay) return;
-      var point = eventPoint(overlay, event);
-      if (!point) return;
-      dragState = {
-        kind: kind,
-        handle: handle,
-        pointerId: event.pointerId,
-        startPoint: point,
-        startBounds: {
-          x: selectedShape.bounds.x,
-          y: selectedShape.bounds.y,
-          width: selectedShape.bounds.width,
-          height: selectedShape.bounds.height
-        }
-      };
-      try {
-        overlay.setPointerCapture(event.pointerId);
-      } catch (_error) {
-        // Pointer capture can be unavailable after a fast re-render; window listeners still cover drag.
-      }
-      window.addEventListener("pointermove", updateDrag);
-      window.addEventListener("pointerup", finishDrag);
-      window.addEventListener("pointercancel", cancelDrag);
-    }
-
-    function updateDrag(event) {
-      if (!dragState || event.pointerId !== dragState.pointerId || !selectedShape) return;
-      var overlay = document.getElementById("selection-overlay");
-      if (!overlay) return;
-      var point = eventPoint(overlay, event);
-      if (!point) return;
-      var dx = point.x - dragState.startPoint.x;
-      var dy = point.y - dragState.startPoint.y;
-      selectedShape.bounds =
-        dragState.kind === "move"
-          ? movedBounds(dragState.startBounds, dx, dy)
-          : resizedBounds(dragState.startBounds, dragState.handle, dx, dy);
-      renderSelectionOverlay();
-    }
-
-    function finishDrag(event) {
-      if (!dragState || event.pointerId !== dragState.pointerId || !selectedShape) return;
-      var nextBounds = selectedShape.bounds;
-      var previousDrag = dragState;
-      dragState = null;
-      detachDragListeners();
-      applyShapeBoundsEdit(previousDrag.startBounds, nextBounds).catch(function (err) {
-        setEditorMessage(err.message, true);
-        selectedShape.bounds = previousDrag.startBounds;
-        renderSelectionOverlay();
-      });
-    }
-
-    function cancelDrag(event) {
-      if (!dragState || event.pointerId !== dragState.pointerId || !selectedShape) return;
-      selectedShape.bounds = dragState.startBounds;
-      dragState = null;
-      detachDragListeners();
-      renderSelectionOverlay();
-    }
-
-    function detachDragListeners() {
-      window.removeEventListener("pointermove", updateDrag);
-      window.removeEventListener("pointerup", finishDrag);
-      window.removeEventListener("pointercancel", cancelDrag);
-    }
-
-    function eventPoint(svg, event) {
-      var matrix = svg.getScreenCTM();
-      if (!matrix) return null;
-      var point = svg.createSVGPoint();
-      point.x = event.clientX;
-      point.y = event.clientY;
-      return point.matrixTransform(matrix.inverse());
-    }
-
-    function movedBounds(bounds, dx, dy) {
-      return {
-        x: bounds.x + dx,
-        y: bounds.y + dy,
-        width: bounds.width,
-        height: bounds.height
-      };
-    }
-
-    function resizedBounds(bounds, handle, dx, dy) {
-      var minSize = 8;
-      var right = bounds.x + bounds.width;
-      var bottom = bounds.y + bounds.height;
-      var next = {
-        x: bounds.x,
-        y: bounds.y,
-        width: bounds.width,
-        height: bounds.height
-      };
-      if (handle === "nw" || handle === "sw") {
-        next.x = Math.min(bounds.x + dx, right - minSize);
-        next.width = right - next.x;
-      }
-      if (handle === "ne" || handle === "se") {
-        next.width = Math.max(minSize, bounds.width + dx);
-      }
-      if (handle === "nw" || handle === "ne") {
-        next.y = Math.min(bounds.y + dy, bottom - minSize);
-        next.height = bottom - next.y;
-      }
-      if (handle === "sw" || handle === "se") {
-        next.height = Math.max(minSize, bounds.height + dy);
-      }
-      return next;
-    }
-
-    function applyShapeBoundsEdit(startBounds, nextBounds) {
-      if (!selectedShape || !selectedShape.handle || !selectedShape.editableTransform) {
-        return Promise.resolve();
-      }
-      var handle = selectedShape.handle;
-      var changed =
-        Math.round(startBounds.x) !== Math.round(nextBounds.x) ||
-        Math.round(startBounds.y) !== Math.round(nextBounds.y) ||
-        Math.round(startBounds.width) !== Math.round(nextBounds.width) ||
-        Math.round(startBounds.height) !== Math.round(nextBounds.height);
-      if (!changed) return Promise.resolve();
-      return postJson("/api/editor/command", {
-        command: {
-          kind: "setShapeTransform",
-          handle: handle,
-          offsetX: Math.round(nextBounds.x * EMU_PER_PIXEL),
-          offsetY: Math.round(nextBounds.y * EMU_PER_PIXEL),
-          width: Math.round(nextBounds.width * EMU_PER_PIXEL),
-          height: Math.round(nextBounds.height * EMU_PER_PIXEL)
-        }
-      }).then(function (data) {
-        if (data) {
-          applyEditorResponse(data);
-          setEditorMessage("Applied", false);
-        }
-      });
-    }
-
-    function addTextBox() {
-      if (activeTextEditor) {
-        commitTextEditor()
-          .then(addTextBox)
-          .catch(function () {});
-        return;
-      }
-      postJson("/api/editor/add-text-box", { slide: currentIndex + 1 })
-        .then(function (data) {
-          applyEditorResponse(data);
-          setEditorMessage("Text box added", false);
-        })
-        .catch(function (err) {
-          setEditorMessage(err.message, true);
-        });
-    }
-
-    function deleteSelectedShape() {
-      if (activeTextEditor || !selectedShape || !selectedShape.handle || !selectedShape.editableDelete) {
-        return;
-      }
-      postJson("/api/editor/command", {
-        command: {
-          kind: "deleteShape",
-          handle: selectedShape.handle
-        }
-      })
-        .then(function (data) {
-          selectedShape = null;
-          selectedShapeKey = null;
-          applyEditorResponse(data);
-          setEditorMessage("Deleted", false);
-        })
-        .catch(function (err) {
-          setEditorMessage(err.message, true);
-        });
-    }
-
-    function isTypingTarget(target) {
-      return target instanceof HTMLInputElement ||
-        target instanceof HTMLTextAreaElement ||
-        target instanceof HTMLSelectElement ||
-        (target instanceof HTMLElement && target.isContentEditable);
-    }
-
-    function duplicateSlide(index) {
-      if (activeTextEditor) {
-        setEditorMessage("Finish text editing before slide operations", true);
-        return;
-      }
-      var slide = slides[index];
-      if (!slide || !slide.handle) {
-        setEditorMessage("Slide handle is unavailable", true);
-        return;
-      }
-      postJson("/api/editor/command", {
-        command: {
-          kind: "duplicateSlide",
-          handle: slide.handle
-        }
-      })
-        .then(function (data) {
-          applyEditorResponse(data, index + 1);
-          setEditorMessage("Duplicated slide", false);
-        })
-        .catch(function (err) {
-          setEditorMessage(err.message, true);
-        });
-    }
-
-    function deleteSlide(index) {
-      if (activeTextEditor) {
-        setEditorMessage("Finish text editing before slide operations", true);
-        return;
-      }
-      if (slideCount <= 1) {
-        setEditorMessage("Cannot delete the last slide", true);
-        return;
-      }
-      var slide = slides[index];
-      if (!slide || !slide.handle) {
-        setEditorMessage("Slide handle is unavailable", true);
-        return;
-      }
-      postJson("/api/editor/command", {
-        command: {
-          kind: "deleteSlide",
-          handle: slide.handle
-        }
-      })
-        .then(function (data) {
-          applyEditorResponse(data, Math.min(index, (data.slides || slides).length - 1));
-          setEditorMessage("Deleted slide", false);
-        })
-        .catch(function (err) {
-          setEditorMessage(err.message, true);
-        });
-    }
-
-    function syncImageReplacementInput() {
-      var input = document.getElementById("image-replacement-input");
-      if (!input) return;
-      var replacement = selectedShape && selectedShape.editableImageReplacement;
-      input.disabled = !replacement;
-      input.value = "";
-      if (replacement) {
-        input.setAttribute("accept", replacement.accept);
-        input.title =
-          "Replace " + replacement.mediaPartPath + " with another " + replacement.contentType + " image";
-      } else {
-        input.removeAttribute("accept");
-        input.title = "Select an image shape to replace its media";
-      }
-    }
-
-    function replaceSelectedImage(file) {
-      if (!selectedShape || !selectedShape.handle || !selectedShape.editableImageReplacement) {
-        setEditorMessage("Select an image shape before replacing media.", true);
-        return Promise.resolve();
-      }
-      if (file.size > ${String(MAX_IMAGE_REPLACEMENT_BYTES)}) {
-        setEditorMessage("Replacement image is too large.", true);
-        return Promise.resolve();
-      }
-      var handle = selectedShape.handle;
-      var input = document.getElementById("image-replacement-input");
-      if (input) input.disabled = true;
-      return file.arrayBuffer()
-        .then(function (buffer) {
-          return postJson("/api/editor/command", {
-            command: {
-              kind: "replaceImage",
-              handle: handle,
-              bytes: Array.from(new Uint8Array(buffer))
-            }
-          });
-        })
-        .then(function (data) {
-          applyEditorResponse(data);
-          setEditorMessage(imageReplacementMessage(data), false);
-        });
-    }
-
-    function imageReplacementMessage(data) {
-      var warnings = (data && data.warnings) || [];
-      var shared = warnings.find(function (warning) {
-        return warning.code === "shared-media-part";
-      });
-      if (!shared) return "Image replaced";
-      return (
-        "Image replaced; shared media part affects " +
-        shared.referenceCount +
-        " pictures: " +
-        shared.mediaPartPath
-      );
-    }
-
-    function postJson(url, payload) {
-      return fetch(url, {
-        method: "POST",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify(payload || {})
-      }).then(function (res) {
-        return res.json().then(function (data) {
-          if (!res.ok) throw new Error(data.error || "request failed");
-          return data;
-        });
-      });
-    }
-
-    function escapeHtmlClient(value) {
-      return String(value)
-        .replace(/&/g, "&amp;")
-        .replace(/</g, "&lt;")
-        .replace(/>/g, "&gt;")
-        .replace(/"/g, "&quot;");
-    }
-
-    renderThumbnails();
-    updateHistory(editorHistory);
-    loadShapeOptions(1);
-
-    document.getElementById("text-run-select").addEventListener("change", syncTextInput);
-    document.getElementById("image-replacement-input").addEventListener("change", function () {
-      var input = document.getElementById("image-replacement-input");
-      var file = input && input.files ? input.files[0] : null;
-      if (!file) return;
-      replaceSelectedImage(file)
-        .catch(function (err) {
-          setEditorMessage(err.message, true);
-        })
-        .finally(function () {
-          syncImageReplacementInput();
-        });
-    });
-    document.getElementById("apply-text-button").addEventListener("click", function () {
-      var option = textRunOptions[Number(document.getElementById("text-run-select").value || "0")];
-      if (!option) return;
-      postJson("/api/editor/command", {
-        command: {
-          kind: "replaceTextRunPlainText",
-          handle: option.handle,
-          text: document.getElementById("text-run-input").value
-        }
-      })
-        .then(function (data) {
-          applyEditorResponse(data);
-          setEditorMessage("Applied", false);
-        })
-        .catch(function (err) {
-          setEditorMessage(err.message, true);
-        });
-    });
-    document.getElementById("add-text-box-button").addEventListener("click", addTextBox);
-    document.getElementById("delete-shape-button").addEventListener("click", deleteSelectedShape);
-    document.getElementById("undo-button").addEventListener("click", function () {
-      postJson("/api/editor/undo")
-        .then(function (data) {
-          applyEditorResponse(data);
-          setEditorMessage("Undone", false);
-        })
-        .catch(function (err) {
-          setEditorMessage(err.message, true);
-        });
-    });
-    document.getElementById("redo-button").addEventListener("click", function () {
-      postJson("/api/editor/redo")
-        .then(function (data) {
-          applyEditorResponse(data);
-          setEditorMessage("Redone", false);
-        })
-        .catch(function (err) {
-          setEditorMessage(err.message, true);
-        });
-    });
-    document.getElementById("save-button").addEventListener("click", function () {
-      postJson("/api/editor/save")
-        .then(function (data) {
-          updateHistory(data.history);
-          setEditorMessage("Saved: " + data.path, false);
-        })
-        .catch(function (err) {
-          setEditorMessage(err.message, true);
-        });
-    });
-
-    // WebSocket for live reload
-    function connect() {
-      var ws = new WebSocket("ws://" + location.host);
-      var status = document.getElementById("status");
-
-      ws.onopen = function () {
-        status.textContent = "Connected";
-        status.className = "";
-      };
-      ws.onclose = function () {
-        status.textContent = "Disconnected - reconnecting...";
-        status.className = "error";
-        setTimeout(connect, 2000);
-      };
-      ws.onmessage = function (event) {
-        var data = JSON.parse(event.data);
-        if (data.type === "rendering") {
-          status.textContent = "Re-rendering...";
-          status.className = "rendering";
-        } else if (data.type === "reload") {
-          status.textContent = "Updating...";
-          status.className = "rendering";
-          location.reload();
-        } else if (data.type === "error") {
-          status.textContent = "Error: " + data.message;
-          status.className = "error";
-        }
-      };
-    }
-    connect();
-
-    // Initial: resize the main SVG
-    (function () {
-      var svg = document.querySelector("#slide-container svg");
-      if (svg) {
-        svg.removeAttribute("width");
-        svg.removeAttribute("height");
-        svg.style.width = "100%";
-        svg.style.height = "auto";
-      }
-    })();
-
-    // Keyboard navigation
-    document.addEventListener("keydown", function (e) {
-      if (e.key === "Delete" && !isTypingTarget(e.target)) {
-        if (selectedShape && selectedShape.editableDelete) {
-          e.preventDefault();
-          deleteSelectedShape();
-        }
-        return;
-      }
-      if (e.key === "ArrowLeft" && currentIndex > 0) selectSlide(currentIndex - 1);
-      if (e.key === "ArrowRight" && currentIndex < slideCount - 1)
-        selectSlide(currentIndex + 1);
-    });
-  </script>
-</body>
-</html>`;
-}
-
-function generateThumbnailsHtml(slides: SlideSvg[]): string {
-  return slides
-    .map(
-      (s, i) =>
-        `<div class="thumbnail${i === 0 ? " active" : ""}" data-index="${i}">` +
-        `<div class="thumb-label"><span class="thumb-title">Slide ${String(s.slideNumber)}</span>` +
-        `<span class="thumb-actions">` +
-        `<button class="thumb-action" data-testid="duplicate-slide-${String(i)}" data-action="duplicate" data-index="${String(i)}" type="button" title="Duplicate slide">D</button>` +
-        `<button class="thumb-action" data-testid="delete-slide-${String(i)}" data-action="delete" data-index="${String(i)}" type="button" title="Delete slide"${slides.length <= 1 ? " disabled" : ""}>X</button>` +
-        `</span></div>` +
-        `<div class="thumb-svg">${s.svg}</div>` +
-        `</div>`,
-    )
-    .join("");
-}
-
-function escapeHtml(str: string): string {
-  return str
-    .replace(/&/g, "&amp;")
-    .replace(/</g, "&lt;")
-    .replace(/>/g, "&gt;")
-    .replace(/"/g, "&quot;");
-}
-
-function jsonForScript(value: unknown): string {
-  return JSON.stringify(value).replace(/</g, "\\u003c");
+  return generateDevEditorHtml({
+    slides,
+    pptxName,
+    emuPerPixel: EMU_PER_PIXEL,
+    maxImageReplacementBytes: MAX_IMAGE_REPLACEMENT_BYTES,
+  });
 }
 
 // --- HTTP API ---
